Fix currency display in grant cards by updating database queries and formatting logic. Ensured grant cards reflect the correct currency from the database instead of hardcoded values. Updated relevant functions and components to maintain consistent currency formatting across the application.
